proje yapısı genel düzenlemeler.

+++ BlogApp.Domain/Entities/Image.cs
@@ -6,7 +6,7 @@ namespace BlogApp.Domain.Entities
     {
         public string Name { get; set; }
         public string? Title { get; set; }
-        public decimal? Size { get; set; }
+        public int? Size { get; set; }
         public string Path { get; set; }
         public string Type { get; set; }
     }
